fixed 'name not appearing on payments page' and some few structural changes on the payment
With the offline mode, when one taps a client name from search, it
carries the person's details to the payment sheet excluding the person's
name.
Also, payment was done offline and was told it s offline so it
was stored. When I came online the client didn't receive the SMS